Moved Writer trait to its own file
