+++ rezasm-source/rezasm-core/src/simulation/mod.rs
@@ -4,3 +4,4 @@ pub mod register;
 pub mod registry;
 pub mod simulator;
 pub mod transform;
+pub mod writer;
